+++ rmw_fastrtps_cpp/include/rmw_fastrtps_cpp/reader_info.hpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef TYPES__WRITER_INFO_HPP_
-#define TYPES__WRITER_INFO_HPP_
+#ifndef RMW_FASTRTPS_CPP__READER_INFO_HPP_
+#define RMW_FASTRTPS_CPP__READER_INFO_HPP_
 
 #include <cassert>
 #include <map>
@@ -23,18 +23,20 @@
 #include <string>
 #include <vector>
 
-#include "fastrtps/participant/Participant.h"
-#include "fastrtps/rtps/builtin/data/WriterProxyData.h"
-#include "fastrtps/rtps/reader/RTPSReader.h"
-
 #include "rcutils/logging_macros.h"
 
+#include "rmw/error_handling.h"
 #include "rmw/rmw.h"
 
-class WriterInfo : public eprosima::fastrtps::ReaderListener
+#include "fastrtps/participant/Participant.h"
+#include "fastrtps/rtps/builtin/data/ReaderProxyData.h"
+#include "fastrtps/rtps/reader/ReaderListener.h"
+#include "fastrtps/rtps/reader/RTPSReader.h"
+
+class ReaderInfo : public eprosima::fastrtps::ReaderListener
 {
 public:
-  WriterInfo(
+  ReaderInfo(
     eprosima::fastrtps::Participant * participant,
     rmw_guard_condition_t * graph_guard_condition)
   : participant_(participant),
@@ -46,9 +48,9 @@ public:
     eprosima::fastrtps::rtps::RTPSReader *,
     const eprosima::fastrtps::CacheChange_t * const change)
   {
-    eprosima::fastrtps::rtps::WriterProxyData proxyData;
+    eprosima::fastrtps::rtps::ReaderProxyData proxyData;
     if (change->kind == ALIVE) {
-      eprosima::fastrtps::CDRMessage_t tempMsg(0);
+      CDRMessage_t tempMsg(0);
       tempMsg.wraps = true;
       tempMsg.msg_endian = change->serializedPayload.encapsulation ==
         PL_CDR_BE ? BIGEND : LITTLEEND;
@@ -59,9 +61,9 @@ public:
         return;
       }
     } else {
-      eprosima::fastrtps::rtps::GUID_t writerGuid;
-      iHandle2GUID(writerGuid, change->instanceHandle);
-      if (!participant_->get_remote_writer_info(writerGuid, proxyData)) {
+      eprosima::fastrtps::rtps::GUID_t readerGuid;
+      iHandle2GUID(readerGuid, change->instanceHandle);
+      if (!participant_->get_remote_reader_info(readerGuid, proxyData)) {
         return;
       }
     }
@@ -112,4 +114,4 @@ public:
   rmw_guard_condition_t * graph_guard_condition_;
 };
 
-#endif  // TYPES__WRITER_INFO_HPP_
+#endif  // RMW_FASTRTPS_CPP__READER_INFO_HPP_
