+++ rmw_fastrtps_cpp/src/rmw_node.cpp
@@ -42,8 +42,8 @@
 #include "fastrtps/rtps/reader/ReaderListener.h"
 #include "fastrtps/rtps/builtin/discovery/endpoint/EDPSimple.h"
 
-#include "identifier.hpp"
-#include "types/custom_participant_info.hpp"
+#include "rmw_fastrtps_cpp/identifier.hpp"
+#include "rmw_fastrtps_cpp/custom_participant_info.hpp"
 
 extern "C"
 {
