+++ rmw_fastrtps_cpp/src/rmw_node_names.cpp
@@ -26,8 +26,8 @@
 
 #include "fastrtps/Domain.h"
 
-#include "identifier.hpp"
-#include "types/custom_participant_info.hpp"
+#include "rmw_fastrtps_cpp/identifier.hpp"
+#include "rmw_fastrtps_cpp/custom_participant_info.hpp"
 
 extern "C"
 {
