+++ rmw_fastrtps_cpp/src/rmw_publish.cpp
@@ -19,7 +19,7 @@
 #include "rmw/error_handling.h"
 #include "rmw/rmw.h"
 
-#include "identifier.hpp"
+#include "rmw_fastrtps_cpp/identifier.hpp"
 #include "ros_message_serialization.hpp"
 #include "types/custom_publisher_info.hpp"
