+++ rmw_fastrtps_cpp/src/rmw_publisher.cpp
@@ -19,10 +19,10 @@
 #include "rmw/rmw.h"
 
 #include "assign_partitions.hpp"
-#include "identifier.hpp"
+#include "rmw_fastrtps_cpp/identifier.hpp"
 #include "namespace_prefix.hpp"
 #include "qos.hpp"
-#include "types/custom_participant_info.hpp"
+#include "rmw_fastrtps_cpp/custom_participant_info.hpp"
 #include "types/custom_publisher_info.hpp"
 #include "type_support_common.hpp"
