+++ rmw_fastrtps_cpp/src/rmw_request.cpp
@@ -23,7 +23,7 @@
 #include "rmw/rmw.h"
 #include "rmw/types.h"
 
-#include "identifier.hpp"
+#include "rmw_fastrtps_cpp/identifier.hpp"
 #include "ros_message_serialization.hpp"
 #include "types/custom_client_info.hpp"
 #include "types/custom_service_info.hpp"
