+++ rmw_fastrtps_cpp/src/rmw_response.cpp
@@ -21,7 +21,7 @@
 #include "rmw/error_handling.h"
 #include "rmw/rmw.h"
 
-#include "identifier.hpp"
+#include "rmw_fastrtps_cpp/identifier.hpp"
 #include "ros_message_serialization.hpp"
 #include "types/custom_client_info.hpp"
 #include "types/custom_service_info.hpp"
