+++ rmw_fastrtps_cpp/src/rmw_service_names_and_types.cpp
@@ -25,9 +25,9 @@
 #include "rmw/names_and_types.h"
 #include "rmw/rmw.h"
 
-#include "identifier.hpp"
+#include "rmw_fastrtps_cpp/identifier.hpp"
 #include "demangle.hpp"
-#include "types/custom_participant_info.hpp"
+#include "rmw_fastrtps_cpp/custom_participant_info.hpp"
 
 extern "C"
 {
